while rebasing...fixed except for a minus sign due to direction of integration.
Conflicts:
	geometry_mod.f95
	laplace_driver.f95
	laplace_system_mod.f95

using trap rule gives an accuracy of 6 digits in the bad region.
using fmm gives a low accuracy that matches the paper. ie they both don't match.

to prevent running out of memory, removing the fmm arrays from get_closeeval_grid.
this is the commit with both the fmm and trap rule.

replaced the fmm with the trapezoid rule. Implements the logic of the paper but is absurd because it's O(N^2).